% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift.R
\name{BD_shift}
\alias{BD_shift}
\title{Assessing the magnitude of BD shifts with 16S rRNA community
data by calculating the beta diversity between unlabeled control
and labeled treatment gradient fraction communities.}
\usage{
BD_shift(physeq, method = "unifrac", weighted = TRUE, fast = TRUE,
  normalized = FALSE, parallel = FALSE)
}
\arguments{
\item{method}{See phyloseq::distance}

\item{weighted}{Weighted Unifrac (if calculating Unifrac)}

\item{fast}{Fast calculation method}

\item{normalized}{Normalized abundances}

\item{parallel}{Calculate in parallel}

\item{df_dist}{phyloseq object}
}
\value{
a data.frame object of weighted mean distances
}
\description{
This function is meant to compare 16S rRNA sequence communities
from many gradient fractions from 2 gradients: a labeled
treatment (eg., 13C-labeled DNA) and its corresponding unlabeled
control. First, the beta-diversity (e.g, Unifrac) is calculated
pairwise between fraction communities. Then, assuming that the
phyloseq \code{sample_data} contains buoyant density information
on each gradient fraction commumnity
(coded in the phyloseq object as 'Buoyant density'), the beta diversity
between each treatment gradient fraction relative to the overlapping
control fractions is set as the weighted mean of beta diversity
values for all overlapping fractions, with the % overlap used for
weighting.
}
\examples{

data(physeq)
# Subsetting phyloseq by Substrate and Day
params = get_treatment_params(physeq, c('Substrate', 'Day'))
params = dplyr::filter(params, Substrate!='12C-Con')
ex = "(Substrate=='12C-Con' & Day=='${Day}') | (Substrate=='${Substrate}' & Day == '${Day}')"
physeq_l = phyloseq_subset(physeq, params, ex)
# Calculating BD_shift on 1 subset (use lapply to process full list)
wmean1 = BD_shift(physeq_l[[1]])
ggplot(wmean1, aes(BD_min.x, wmean_dist)) +
   geom_point()

# Calculating BD_shift on all subsets
lapply(physeq_l, BD_shift)

}

