% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_atom_excess.R
\name{calc_atom_excess}
\alias{calc_atom_excess}
\title{Calculate atom fraction excess}
\usage{
calc_atom_excess(Mlab, Mlight, Mheavymax, isotope = "13C")
}
\arguments{
\item{Mlab}{The molecular wight of labeled DNA}

\item{Mlight}{The molecular wight of unlabeled DNA}

\item{Mheavymax}{The theoretical maximum molecular weight of fully-labeled DNA}

\item{isotope}{The isotope for which the DNA is labeled with ('13C' or '18O')}
}
\value{
numeric value: atom fraction excess (A)
}
\description{
See Hungate et al., 2015 for more details
}
