% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_atom_excess.R
\name{calc_Gi}
\alias{calc_Gi}
\title{Calculate G+C from unlabeled buoyant density}
\usage{
calc_Gi(Wlight)
}
\arguments{
\item{Wlight}{A vector with >=1 weighted mean BD from
'light' gradient fractions}
}
\value{
numeric value (fractional G+C; Gi)
}
\description{
See Hungate et al., 2015 for more details
}
