% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_atom_excess.R
\name{calc_Mheavymax}
\alias{calc_Mheavymax}
\title{Calculate the theoretical maximum molecular weight of fully-labeled DNA}
\usage{
calc_Mheavymax(Mlight, isotope = "13C", Gi = NA)
}
\arguments{
\item{Mlight}{The molecular wight of unlabeled DNA}

\item{isotope}{The isotope for which the DNA is labeled with ('13C' or '18O')}

\item{Gi}{The G+C content of unlabeled DNA}
}
\value{
numeric value: maximum molecular weight of fully-labeled DNA
}
\description{
See Hungate et al., 2015 for more details
}
