% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{data-physeq_rep3}
\alias{data-physeq_rep3}
\alias{physeq_rep3}
\title{(Data) A simulated HTS-SIP dataset}
\description{
6 gradients:
\itemize{
  \item{*}{12C-control, replicate 1}
  \item{*}{12C-control, replicate 2}
  \item{*}{12C-control, replicate 3}
  \item{*}{13C-treatment, replicate 1}
  \item{*}{13C-treatment, replicate 2}
  \item{*}{13C-treatment, replicate 3}
  }
}
\keyword{data}
