% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{data-physeq_rep3_qPCR}
\alias{data-physeq_rep3_qPCR}
\alias{physeq_rep3_qPCR}
\title{(Data) qPCR data associated with the physeq_rep3 HTS-SIP dataset}
\description{
The dataset contains simulated 16S rRNA copies for each gradient
fraction of each gradient (measured via qPCR).
}
\seealso{
\code{physeq_rep3}
}
\keyword{data}
