% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{data-physeq_S2D2}
\alias{data-physeq_S2D2}
\alias{physeq_S2D2}
\title{(Data) A subset of full HTS-SIP dataset (Substrates=2, Days=2)}
\description{
1 of the 2 'substrates' is the 12C-control;
the other is the 13C-labeled substrate.
}
\keyword{data}
