% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{data-physeq_S2D2_l}
\alias{data-physeq_S2D2_l}
\alias{physeq_S2D2_l}
\title{(Data) A subset of full HTS-SIP dataset (Substrates=2, Days=2)}
\description{
1 of the 2 'substrates' is the 12C-control;
the other is the 13C-labeled substrate.
}
\details{
The dataset has been parsed into a list of corresponding
13C-treatment vs 12C-control comparisons. Each comparison
is of all gradient fraction samples of the treatment vs
the gradient fraction samples of the control.
}
\keyword{data}
