% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_l2fc.R
\name{DESeq2_l2fc}
\alias{DESeq2_l2fc}
\title{Format phyloseq & call DESeq2}
\usage{
DESeq2_l2fc(physeq, density_min, density_max, design, l2fc_threshold = 0.25,
  sparsity_threshold = 0.25, sparsity_apply = "all", ...)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{density_min}{Minimum buoyant density of the 'heavy' gradient fractions}

\item{density_max}{Maximum buoyant density of the 'heavy' gradient fractions}

\item{design}{\code{design} parameter used for DESeq2 analysis.
See \code{DESeq2::DESeq} for more details.}

\item{l2fc_threshold}{log2 fold change (l2fc) values must be significantly above this
threshold in order to reject the hypothesis of equal counts.}

\item{sparsity_threshold}{All OTUs observed in less than this portion (fraction: 0-1)
of gradient fraction samples are pruned. A a form of indepedent filtering,
The sparsity cutoff with the most rejected hypotheses is used.}

\item{sparsity_apply}{Apply sparsity threshold to all gradient fraction samples ('all')
or just heavy fraction samples ('heavy')}
}
\value{
dataframe of HRSIP results
}
\description{
Conducts the actual trimming of the phyloseq dataset & calling of DESeq2
}
\examples{
data(physeq)
df_l2fc = DESeq2_l2fc(physeq, density_min=1.71, density_max=1.75, design=~Substrate)
head(df_l2fc)

}

