% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_l2fc.R
\name{DESeq2_l2fc_multi}
\alias{DESeq2_l2fc_multi}
\title{Calculating l2fc for multiple sparsity thresholds}
\usage{
DESeq2_l2fc_multi(physeq, density_windows, design, l2fc_threshold = 0.25,
  sparsity_threshold = seq(0, 0.5, 0.1), sparsity_apply = "all",
  parallel = FALSE, ...)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{density_windows}{2-column data.frame with min & max buoyant densities of 'heavy' gradient
fraction windows (1 window per row). The gradient fractions in each window are used to calculate
log2 fold change values.}

\item{design}{\code{design} parameter used for DESeq2 analysis.
See \code{DESeq2::DESeq} for more details.}

\item{l2fc_threshold}{log2 fold change (l2fc) values must be significantly above this
threshold in order to reject the hypothesis of equal counts.}

\item{sparsity_threshold}{All OTUs observed in less than this portion (fraction: 0-1)
of gradient fraction samples are pruned. A a form of indepedent filtering,
The sparsity cutoff with the most rejected hypotheses is used.}

\item{sparsity_apply}{Apply sparsity threshold to all gradient fraction samples ('all')
or just heavy fraction samples ('heavy')}

\item{parallel}{Run in parallel. See plyr::mdply for more information (.parallel flag).}
}
\value{
dataframe of HRSIP results
}
\description{
Calls \code{DESeq2_l2fc} with different sparsity thresholds
and uses the l2fc results of the sparsity threshold that produced
the most rejected hypotheses
}
\examples{
data(physeq)
df_l2fc = DESeq2_l2fc_multi(physeq, density_min=1.71, density_max=1.75, design=~Substrate)
head(df_l2fc)

}

