% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_l2fc_sub.R
\name{DESeq2_l2fc_multi_sub}
\alias{DESeq2_l2fc_multi_sub}
\title{\code{DESeq2_l2fc_multi}: subset comparison}
\usage{
DESeq2_l2fc_multi_sub(params, ex, physeq, density_windows,
  design = ~Substrate, l2fc_threshold = 0.25, sparsity_threshold = seq(0,
  0.3, 0.1), sparsity_apply = "all", parallel = parallel)
}
\arguments{
\item{params}{data.frame of parameters supplies to \code{ex}}

\item{ex}{Expression for subsetting the phyloseq object}

\item{physeq}{Phyloseq object}

\item{density_windows}{Min/max windows for selecting gradient fractions for l2fc calculations.}

\item{design}{\code{design} parameter used for DESeq2 analysis.
See \code{DESeq2::DESeq} for more details.}

\item{l2fc_threshold}{log2 fold change (l2fc) values must be significantly above this
threshold in order to reject the hypothesis of equal counts.}

\item{sparsity_threshold}{All OTUs observed in less than this portion (fraction: 0-1)
of gradient fraction samples are pruned. A a form of indepedent filtering,
The sparsity cutoff with the most rejected hypotheses is used.}

\item{sparsity_apply}{Apply sparsity threshold to all gradient fraction samples ('all')
or just heavy fraction samples ('heavy')}
}
\value{
dataframe of HRSIP results
}
\description{
Calls \code{DESeq2_l2fc_multi} with subsets of the dataset:
(eg., just gradient fractions from 1 treatment & its matching
control gradient samples)
}
\details{
The phyloseq object subsetting is performed with an expression
that generates boolean values (eg., 1:10 < 3). Since each subset
requires a specific expression, the user provides a 'generalized'
expression in which specific values are inserted into the specific
expression for each subset.
For example: expression="(Substrate=='12C-Con') |
(Substrate=='${Substrate}')"
The "${Substrate}" value is changed for each subset, and the values
used for the parameter are set by the \code{params} argument.
}
\examples{
data(physeq)
params = get_treatment_params(physeq, c('Substrate', 'Day'), "Substrate != '12C-Con'")
params_l = apply(params, 1, as.list)
ex = "(Substrate=='12C-Con' & Day=='${Day}') | (Substrate=='${Substrate}' & Day == '${Day}')"
df_l2fc = plyr::ldply(params_l, DESeq2_l2fc_multi_sub, ex=ex, physeq=physeq)
head(df_l2fc)

}

