% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringterpolate.R
\name{evaluate_matches}
\alias{evaluate_matches}
\title{Evaluate String Interpolation Matches}
\usage{
evaluate_matches(matches, env)
}
\arguments{
\item{matches}{Match data}

\item{env}{The environment in which to evaluate the expressions.}
}
\value{
A character vector of replacement strings.
}
\description{
The expression part of string interpolation matches are evaluated in a
specified environment and formatted for replacement in the original string.
}
