% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSIP.R
\name{filter_l2fc}
\alias{filter_l2fc}
\title{Filter l2fc table to 'best' sparsity cutoffs & density windows}
\usage{
filter_l2fc(df_l2fc, padj_cutoff = 0.1, padj_method = "BH")
}
\arguments{
\item{df_l2fc}{data.frame of log2 fold change values}

\item{padj_cutoff}{Adjusted p-value cutoff for rejecting the null hypothesis
that l2fc values were not greater than the l2fc_threshold.}

\item{padj_method}{Method for global p-value adjustment (See \code{p.adjust})}
}
\description{
\code{filter_l2fc} filters a l2fc table to 'best' sparsity cutoffs & density windows.
}

