% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift.R
\name{fraction_overlap}
\alias{fraction_overlap}
\title{Calculate the BD range overlap of gradient fractions}
\usage{
fraction_overlap(metadata)
}
\arguments{
\item{metadata}{Metdata data.frame object. See \code{format_metadata()}.}
}
\value{
a data.frame object of metadata with fraction BD overlaps
}
\description{
Calculate the BD range overlap of gradient fractions
}
\examples{
\dontrun{
data(physeq_S2D2)
ex = "Substrate=='12C-Con'"
metadata = HTSSIP:::format_metadata(physeq_S2D2, ex)
m = HTSSIP:::fraction_overlap(metadata)
head(m)
}

}
