% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSIP.R
\name{HRSIP}
\alias{HRSIP}
\title{HR-SIP analysis}
\usage{
HRSIP(physeq, density_windows, design, sparsity_threshold = seq(0, 0.3, 0.1),
  sparsity_apply = "all", l2fc_threshold = 0.25, pairwise_expr = NULL,
  pairwise_params = NULL, padj_cutoff = 0.1, padj_method = "BH",
  parallel = FALSE, ...)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{design}{\code{design} parameter used for DESeq2 analysis.
See \code{DESeq2::DESeq} for more details.}

\item{sparsity_threshold}{All OTUs observed in less than this portion (fraction: 0-1)
of gradient fraction samples are pruned. A a form of indepedent filtering,
The sparsity cutoff with the most rejected hypotheses is used.}

\item{sparsity_apply}{Apply sparsity threshold to all gradient fraction samples ('all')
or just heavy fraction samples ('heavy')}

\item{l2fc_threshold}{log2 fold change (l2fc) values must be significantly above this
threshold in order to reject the hypothesis of equal counts.}

\item{padj_cutoff}{Adjusted p-value cutoff for rejecting the null hypothesis
that l2fc values were not greater than the l2fc_threshold.}

\item{padj_method}{Method for global p-value adjustment (See \code{p.adjust})}

\item{parallel}{Run in parallel. See plyr::mdply for more information (.parallel flag).}
}
\description{
\code{HRSIP} conducts high resolution stable isotope probing (HR-SIP)
  analysis on phyloseq object.
}
\examples{
data(physeq)
# This will compare all 13C treatments to the 12C controls (all time points combined)
density_windows = data.frame(start=c(1.70), end=c(1.75))
df_l2fc = HRSIP(physeq, density_windows=density_windows, design=~Substrate)
head(df_l2fc)

# This will conduct individual comparisons between 12C controls & 13C treatments (by time point)
## Get a listing of the parameters used for subsetting the phyloseq object for each comparison
params = get_treatment_params(physeq, c('Substrate', 'Day'), "Substrate != '12C-Con'")
## Make an expression for subsetting the phyloseq object.
## The values in braces will be replaced by values in the param.
## The first "()" is selecting the control; the second "()" is selecting a treatment.
ex = "(Substrate=='12C-Con' & Day=='${Day}') | (Substrate=='${Substrate}' & Day == '${Day}')"
## HRSIP run on each comparison
density_windows = data.frame(start=c(1.70), end=c(1.75))
df_l2fc = HRSIP(physeq, pairwise_expr=ex, pairwise_params=params, density_windows=density_windows, design=~Substrate)
head(df_l2fc)

# Same as above, using multiple buoyant density windows (MW-HR-SIP)
density_windows = data.frame(start=c(1.70, 1.72), end=c(1.73, 1.75))
df_l2fc = HRSIP(physeq, pairwise_expr=ex, pairwise_params=params, density_windows=density_windows, design=~Substrate)
head(df_l2fc)

# Same as above, but running in parallel
library(doParallel)
ncores=4
registerDoParallel(ncores)
df_l2fc = HRSIP(physeq, pairwise_expr=ex, pairwise_params=params, density_windows=density_windows, design=~Substrate, .parallel=TRUE)
head(df_l2fc)

}

