% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\docType{package}
\name{HTSSIP}
\alias{HTSSIP}
\alias{HTSSIP-package}
\title{HTSSIP: A package for analyzing high throughput sequence stable isotope probing data}
\description{
HTSSIP: A package for analyzing high throughput sequence stable isotope probing data
}
\section{HTSSIP functions}{

HTSSIP func1
}

