% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringterpolate.R
\name{match_placeholders}
\alias{match_placeholders}
\title{Match Expression Placeholders for String Interpolation}
\usage{
match_placeholders(string)
}
\arguments{
\item{string}{character: The string to be interpolated.}
}
\value{
list containing \code{indices} (regex match data) and \code{matches},
  the string representations of matched expressions.
}
\description{
Given a character string a set of expression placeholders are matched. They
are of the form \code{${...}} or optionally \code{$[f]{...}} where \code{f}
is a valid format for \code{\link{sprintf}}.
}
