% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_trans.R
\name{OTU_qPCR_trans}
\alias{OTU_qPCR_trans}
\title{Transform OTU counts based on qPCR data}
\usage{
OTU_qPCR_trans(physeq, qPCR)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{qPCR}{A list of qPCR data from \code{qPCR_sim()}}
}
\value{
A phyloseq object with transformed OTU counts
}
\description{
Transform OTU counts based on qPCR data
}
\examples{
# qPCR data simulation
## making functions for simulating values
control_mean_fun = function(x) dnorm(x, mean=1.70, sd=0.01) * 1e8
control_sd_fun = function(x) control_mean_fun(x) / 3
treat_mean_fun = function(x) dnorm(x, mean=1.75, sd=0.01) * 1e8
treat_sd_fun = function(x) treat_mean_fun(x) / 3
## simulating qPCR values
qPCR = qPCR_sim(physeq_S2D2,
                control_expr='Substrate=="12C-Con"',
                control_mean_fun=control_mean_fun,
                control_sd_fun=control_sd_fun,
                treat_mean_fun=treat_mean_fun,
                treat_sd_fun=treat_sd_fun)
## transforming OTU values
OTU_qPCR_trans(physeq, qPCR)

}

