% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift.R
\name{overlap_wmean_dist}
\alias{overlap_wmean_dist}
\title{Calculating weighted mean beta-diversities of overlapping gradient fractions.}
\usage{
overlap_wmean_dist(df_dist)
}
\arguments{
\item{df_dist}{Filtered distance matrix in data.frame format.
See \code{parse_dist()}}
}
\value{
a data.frame object of weighted mean distances
}
\description{
Calculating weighted mean beta-diversities of overlapping gradient fractions.
}
\examples{
\dontrun{
data(physeq_S2D2)
physeq_S2D2_d = phyloseq::distance(physeq_S2D2,
                             method='unifrac',
                             weighted=TRUE,
                             fast=TRUE,
                             normalized=FALSE)
physeq_S2D2_d = parse_dist(physeq_S2D2_d)
wmean = overlap_wmean_dist(physeq_S2D2_d)
head(wmean)
}

}

