% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift.R
\name{parse_dist}
\alias{parse_dist}
\title{Filtering out non-relevant distances in distance matrix}
\usage{
parse_dist(d)
}
\arguments{
\item{metadata}{Metdata data.frame object. See \code{format_metadata()}.}
}
\value{
a data.frame object of metadata with fraction BD overlaps
}
\description{
Filtering out non-relevant distances in distance matrix
}
\examples{

data(physeq)
physeq_d = phyloseq::distance(physeq,
                             method='unifrac',
                             weighted=TRUE,
                             fast=TRUE,
                             normalized=FALSE)
physeq_d = parse_dist(physeq_d)
head(physeq_d)

}

