% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift.R
\name{perc_overlap}
\alias{perc_overlap}
\title{Calculate the percent overlap between two ranges (x & y).}
\usage{
perc_overlap(x.start, x.end, y.start, y.end)
}
\arguments{
\item{x.start}{The start value for Range X}

\item{x.end}{The end value for Range X}

\item{y.start}{The start value for Range Y}

\item{y.end}{The end value for Range Y}
}
\value{
the percent overlap of the ranges
}
\description{
The fraction of overlap is relative to Range X (see examples).
}
\examples{
\dontrun{
x = HTSSIP:::perc_overlap(0, 1, 0, 0.5)
stopifnot(x == 50)
x = HTSSIP:::perc_overlap(0, 0.5, 0, 1)
stopifnot(x == 100)
}

}
