% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSIP.R
\name{phyloseq2df}
\alias{phyloseq2df}
\title{phyloseq data object conversion to data.frame}
\usage{
phyloseq2df(physeq, table_func)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{table_func}{See \code{Phyloseq::phyloseq-class} for options}
}
\value{
dataframe
}
\description{
Conducts conversion of 1 of the data objects
in a phyloseq object (eg., tax_table) to a dataframe
}
\examples{
data(physeq)
df_otu = phyloseq2df(physeq, table_func=otu_table)
head(df_otu)
}

