% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift_ordinations.R
\name{phyloseq_list_ord_dfs}
\alias{phyloseq_list_ord_dfs}
\title{Converting ordination objects to data.frames}
\usage{
phyloseq_list_ord_dfs(physeq_l, physeq_l_ords)
}
\arguments{
\item{physeq_l}{A list of phyloseq objects}

\item{physeq_l_ords}{A list of ordination objects}
}
\value{
List of data.frame objects
}
\description{
For each ordination object in a list, converts to a data.frame
for easy plotting with ggplot
}
\examples{
data(physeq_l)
physeq_l_d = physeq_list_betaDiv(physeq_l)
physeq_l_d_ord = physeq_list_ord(physeq_l, physeq_l_d)
physeq_l_d_ord_df = phyloseq_list_ord_dfs(physeq_l, physeq_l_d_ord)

}

