% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{physeq_format}
\alias{physeq_format}
\title{Checking format of phyloseq object for HTSSIP compatibility}
\usage{
physeq_format(physeq)
}
\arguments{
\item{physeq}{Phyloseq object}
}
\value{

}
\description{
Checking format of phyloseq object for HTSSIP compatibility
}
\examples{

# should be formatted for HTSSIP
data(physeq_S2D2)
physeq_format(physeq_S2D2)

# should NOT be correctly formatted for HTSSIP
library(phyloseq)
data(GlobalPatterns)
tryCatch(
 physeq_format(GlobalPatterns),
 function(e) e
 )

}

