% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_ordinations.R
\name{physeq_list_ord}
\alias{physeq_list_ord}
\title{calculating beta diversity for a list of phyloseq objects}
\usage{
physeq_list_ord(physeq_l, physeq_l_d, ord_method = "NMDS")
}
\arguments{
\item{physeq_l}{A list of phyloseq objects}

\item{physeq_l_d}{A list of dist objects}

\item{ord_method}{See phyloseq::ordinate}
}
\value{
List of ordination objects
}
\description{
For each dist object in a list, calculates ordinations
}
\examples{
data(physeq_l)
physeq_l_d = physeq_list_betaDiv(physeq_l)
physeq_l_d_ord = physeq_list_ord(physeq_l, physeq_l_d)

}

