% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{qPCR_sim}
\alias{qPCR_sim}
\title{Simulate qPCR values}
\usage{
qPCR_sim(physeq, control_mean_fun, control_sd_fun, treat_mean_fun, treat_sd_fun,
  n_tech_rep = 3, control_expr = NULL, ...)
}
\arguments{
\item{physeq}{Object of class "phyloseq"}

\item{control_mean_fun}{Function used for simulating the qPCR normal distribution mean
for control samples.}

\item{control_sd_fun}{Function used for simulating the qPCR normal distribution
standard deviation for control samples.}

\item{treat_mean_fun}{Function used for simulating the qPCR normal distribution mean
for treatment samples.}

\item{treat_sd_fun}{Function used for simulating the qPCR normal distribution
standard deviation for treatment samples.}

\item{n_tech_rep}{Number of technical replicates.}

\item{control_expr}{Expression used to identify control samples based on sample_data.}
}
\value{
data.frame of qPCR values
}
\description{
Simulate qPCR values
}
\examples{
# making functions for simulating values
control_mean_fun = function(x) dnorm(x, mean=1.70, sd=0.01) * 1e8
control_sd_fun = function(x) control_mean_fun(x) / 3
treat_mean_fun = function(x) dnorm(x, mean=1.75, sd=0.01) * 1e8
treat_sd_fun = function(x) treat_mean_fun(x) / 3
# simulating qPCR values
df_qPCR = qPCR_sim(physeq_S2D2,
                control_expr='Substrate=="12C-Con"',
                control_mean_fun=control_mean_fun,
                control_sd_fun=control_sd_fun,
                treat_mean_fun=treat_mean_fun,
                treat_sd_fun=treat_sd_fun)

}

