% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_atom_excess.R
\name{qSIP_atom_excess}
\alias{qSIP_atom_excess}
\title{Calculate atom fraction excess using q-SIP method}
\usage{
qSIP_atom_excess(physeq, control_expr, treatment_rep = NULL,
  isotope = "13C", df_OTU_W = NULL)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{control_expr}{Expression used to identify control samples based on sample_data.}

\item{treatment_rep}{Which column in the phyloseq sample data designates
replicate treatments}

\item{isotope}{The isotope for which the DNA is labeled with ('13C' or '18O')}

\item{df_OTU_W}{Keep NULL}
}
\value{
A list of 2 data.frame objects. 'W' contains
the weighted mean buoyant density (W) values for each OTU in
each treatment/control. 'A' contains the atom fraction excess
values for each OTU. For the 'A' table, the 'Z' column is buoyant
density shift, and the 'A' column is atom fraction excess.
}
\description{
Calculate atom fraction excess using q-SIP method
}
\examples{
# qPCR data simulation
control_mean_fun = function(x) dnorm(x, mean=1.70, sd=0.01) * 1e8
control_sd_fun = function(x) control_mean_fun(x) / 3
treat_mean_fun = function(x) dnorm(x, mean=1.75, sd=0.01) * 1e8
treat_sd_fun = function(x) treat_mean_fun(x) / 3

physeq = physeq_S2D2_l[[1]]
qPCR = qPCR_sim(physeq,
                control_expr='Substrate=="12C-Con"',
                control_mean_fun=control_mean_fun,
                control_sd_fun=control_sd_fun,
                treat_mean_fun=treat_mean_fun,
                treat_sd_fun=treat_sd_fun)
# OTU table transformation
physeq_t = OTU_qPCR_trans(physeq, qPCR)

#BD shift (Z) & atom excess (A)
atomX = qSIP_atom_excess(physeq_t,
                        control_expr='Substrate=="12C-Con"',
                        treatment_rep='Microcosm_replicate')

}

