% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_atom_excess.R
\name{qSIP_bootstrap}
\alias{qSIP_bootstrap}
\title{Calculate bootstrap CI for atom fraction excess using q-SIP method}
\usage{
qSIP_bootstrap(atomX, isotope = "13C", n_sample = c(3, 3), n_boot = 10,
  parallel = FALSE, a = 0.1)
}
\arguments{
\item{atomX}{A list object created by \code{qSIP_atom_excess()}}

\item{isotope}{The isotope for which the DNA is labeled with ('13C' or '18O')}

\item{n_sample}{A vector of length 2.
The sample size for data resampling (with replacement) for 1) control samples
and 2) treatment samples.}

\item{n_boot}{Number of bootstrap replicates.}

\item{parallel}{Parallel processing. See \code{.parallel} option in
\code{dplyr::mdply()} for more details.}

\item{a}{A numeric value. The alpha for calculating confidence intervals.}
}
\value{
A data.frame of atom fraction excess values (A) and
atom fraction excess confidence intervals.
}
\description{
Calculate bootstrap CI for atom fraction excess using q-SIP method
}
\examples{
# qPCR data simulation
control_mean_fun = function(x) dnorm(x, mean=1.70, sd=0.01) * 1e8
control_sd_fun = function(x) control_mean_fun(x) / 3
treat_mean_fun = function(x) dnorm(x, mean=1.75, sd=0.01) * 1e8
treat_sd_fun = function(x) treat_mean_fun(x) / 3

physeq = physeq_S2D2_l[[1]]
qPCR = qPCR_sim(physeq,
                control_expr='Substrate=="12C-Con"',
                control_mean_fun=control_mean_fun,
                control_sd_fun=control_sd_fun,
                treat_mean_fun=treat_mean_fun,
                treat_sd_fun=treat_sd_fun)
# OTU table transformation
physeq_t = OTU_qPCR_trans(physeq, qPCR)

#BD shift (Z) & atom excess (A)
atomX = qSIP_atom_excess(physeq_t,
                        control_expr='Substrate=="12C-Con"',
                        treatment_rep='Microcosm_replicate')

# bootstrapping in parallel
doParallel::registerDoParallel(8)
df_atomX_boot = qSIP_bootstrap(atomX, parallel=TRUE)
head(df_atomX_boot)

}

