% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift_ordinations.R
\name{SIP_betaDiv_ord}
\alias{SIP_betaDiv_ord}
\title{Calculating & plotting beta diversity for a list of phyloseq objects}
\usage{
SIP_betaDiv_ord(physeq_l, method = "unifrac", weighted = TRUE,
  fast = TRUE, normalized = TRUE, parallel = FALSE)
}
\arguments{
\item{physeq_l}{A list of phyloseq objects}

\item{method}{See phyloseq::distance}

\item{weighted}{Weighted Unifrac (if calculating Unifrac)}

\item{fast}{Fast calculation method}

\item{normalized}{Normalized abundances}

\item{parallel}{Calculate in parallel}
}
\value{
List of dist objects
}
\description{
For each phyloseq object in a list, calculates beta-diversity
between all samples using the phyloseq::distance function.
}
\examples{
data(physeq_l)
physeq_l_p = SIP_betaDiv(physeq_l)
do.call(gridExtra::grid.arrange, c(physeq_l_p))

}

