% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringterpolate.R
\name{stringterpolate}
\alias{stringterpolate}
\title{String Interpolation}
\usage{
stringterpolate(string, env = parent.frame())
}
\arguments{
\item{string}{A template character string.}

\item{env}{The environment in which to evaluate the expressions.}
}
\value{
An interpolated character string.
}
\description{
String interpolation is a useful way of specifying a character string which
depends on values in a certain environment. It allows for string creation
which is easier to read and write when compared to using e.g.
\code{\link{paste}} or \code{\link{sprintf}}. The (template) string can
include expression placeholders of the form \code{${expression}} or
\code{$[format]{expression}}, where expressions are valid R expressions that
can be evaluated in the given environment, and \code{format} is a format
specification valid for use with \code{\link{sprintf}}.
}
