% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSIP_trans.R
\name{tss}
\alias{tss}
\title{Total sum scaling}
\usage{
tss(x, MARGIN = 2, na.rm = FALSE)
}
\arguments{
\item{x}{data.frame of numeric values}

\item{MARGIN}{table margin (1=rows, 2=columns)}

\item{na.rm}{remove NAs?}
}
\value{
data.frame of qPCR values
}
\description{
Total sum scaling
}
\examples{
# making functions for simulating values
df = data.frame(1:5, 5:9)
df_t = tss(df)
apply(df_t, 2, sum)

}
