% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRSIP.R
\name{HRSIP}
\alias{HRSIP}
\title{HR-SIP analysis}
\usage{
HRSIP(physeq, sparsity_threshold, density_min, density_max, design,
  l2fc_threshold = 0.25, sparsity_apply = "all", ...)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{sparsity_threshold}{All OTUs observed in less than this portion (fraction: 0-1)
of gradient fraction samples are pruned}

\item{l2fc_threshold}{DESeq2 log2 fold change threshold}

\item{sparsity_apply}{Apply sparsity threshold to all gradient fraction samples ('all')
or just heavy fraction samples ('heavy')}
}
\value{
dataframe of HRSIP results
}
\description{
\code{HRSIP} conducts HR-SIP analysis on phyloseq object
}
\details{
blah blah blah
}
\examples{
HRSIP()

}

