% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSSIP.R
\docType{package}
\name{HTSSIP}
\alias{HTSSIP}
\alias{HTSSIP-package}
\title{HTSSIP: analyzing high throughput sequence data from nucleotide stable isotope probing experiments}
\description{
HTSSIP provides a toolset for reproducibly analyzing HTS-SIP data.
HTS-SIP data is the combination of nucleotide stable isotope probing (DNA- & RNA-SIP)
and high throughput sequence data (e.g., MiSeq of 16S rRNA amplicons).
}
\details{
To learn more about HTSSIP, start with the vignettes:
\code{browseVignettes(package = "HTSSIP")}
}
