% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_shift.R
\name{parse_dist}
\alias{parse_dist}
\title{Filtering out non-relevant distances in distance matrix}
\usage{
parse_dist(d)
}
\arguments{
\item{d}{a distance matrix object}
}
\value{
a data.frame object of metadata with fraction BD overlaps
}
\description{
Filtering out non-relevant distances in distance matrix
}
\examples{
\dontrun{
data(physeq_S2D2)
physeq_S2D2_d = phyloseq::distance(physeq_S2D2,
                             method='unifrac',
                             weighted=TRUE,
                             fast=TRUE,
                             normalized=FALSE)
physeq_S2D2_d = HTSSIP:::parse_dist(physeq_S2D2_d)
head(physeq_S2D2_d)
}

}
